+++ packages/transaction-pay-controller/src/actions/update-payment-token.ts
@@ -1,19 +1,13 @@
 import { createModuleLogger } from '@metamask/utils';
 import type { Hex } from '@metamask/utils';
-import { BigNumber } from 'bignumber.js';
 
 import type { TransactionPayControllerMessenger } from '..';
 import { projectLogger } from '../logger';
 import type {
-  TransactionPaymentToken,
   UpdatePaymentTokenRequest,
   UpdateTransactionDataCallback,
 } from '../types';
-import {
-  getTokenBalance,
-  getTokenFiatRate,
-  getTokenInfo,
-} from '../utils/token';
+import { buildTokenData } from '../utils/token';
 import { getTransaction } from '../utils/transaction';
 
 const log = createModuleLogger(projectLogger, 'update-payment-token');
@@ -42,7 +36,7 @@ export function updatePaymentToken(
     throw new Error('Transaction not found');
   }
 
-  const paymentToken = getPaymentToken({
+  const paymentToken = buildTokenData({
     chainId,
     from: transaction?.txParams.from as Hex,
     messenger,
@@ -58,64 +52,4 @@ export function updatePaymentToken(
   updateTransactionData(transactionId, (data) => {
     data.paymentToken = paymentToken;
   });
-}
-
-/**
- * Generate the full payment token data from a token address and chain ID.
- *
- * @param request - The payment token request parameters.
- * @param request.chainId - The chain ID.
- * @param request.from - The address to get the token balance for.
- * @param request.messenger - The transaction pay controller messenger.
- * @param request.tokenAddress - The token address.
- * @returns The payment token or undefined if the token data could not be retrieved.
- */
-function getPaymentToken({
-  chainId,
-  from,
-  messenger,
-  tokenAddress,
-}: {
-  chainId: Hex;
-  from: Hex;
-  messenger: TransactionPayControllerMessenger;
-  tokenAddress: Hex;
-}): TransactionPaymentToken | undefined {
-  const { decimals, symbol } =
-    getTokenInfo(messenger, tokenAddress, chainId) ?? {};
-
-  if (decimals === undefined || !symbol) {
-    return undefined;
-  }
-
-  const tokenFiatRate = getTokenFiatRate(messenger, tokenAddress, chainId);
-
-  if (tokenFiatRate === undefined) {
-    return undefined;
-  }
-
-  const balance = getTokenBalance(messenger, from, chainId, tokenAddress);
-  const balanceRawValue = new BigNumber(balance);
-  const balanceHumanValue = new BigNumber(balance).shiftedBy(-decimals);
-  const balanceRaw = balanceRawValue.toFixed(0);
-  const balanceHuman = balanceHumanValue.toString(10);
-
-  const balanceFiat = balanceHumanValue
-    .multipliedBy(tokenFiatRate.fiatRate)
-    .toString(10);
-
-  const balanceUsd = balanceHumanValue
-    .multipliedBy(tokenFiatRate.usdRate)
-    .toString(10);
-
-  return {
-    address: tokenAddress,
-    balanceFiat,
-    balanceHuman,
-    balanceRaw,
-    balanceUsd,
-    chainId,
-    decimals,
-    symbol,
-  };
 }